[Cast] refactor kernels
