+++ include/flexflow/ops/kernels/cast_kernels.h
@@ -1,0 +1,44 @@
+#ifndef _FLEXFLOW_OPS_KERNELS_CAST_KERNELS_H
+#define _FLEXFLOW_OPS_KERNELS_CAST_KERNELS_H
+
+#include "flexflow/device.h"
+#include "flexflow/fftype.h"
+#include "flexflow/op_meta.h"
+
+namespace FlexFlow {
+
+class CastMeta : public OpMeta {
+public:
+  CastMeta(FFHandler handle);
+  DataType input_data_type, output_data_type;
+};
+
+namespace Kernels {
+namespace Cast {
+template <typename IDT, typename ODT>
+ void forward_kernel_wrapper(const IDT *input_ptr,
+                                    ODT *output_ptr,
+                                    size_t volume);
+
+template <typename IDT, typename ODT>
+ void
+    backward_kernel_wrapper(const IDT *src_ptr, ODT *dst_ptr, size_t volume);
+
+namespace Internal {
+
+template <typename IDT, typename ODT>
+ void forward_kernel(const IDT *input_ptr,
+                             ODT *output_ptr,
+                             size_t volume,
+                             ffStream_t stream);
+template <typename IDT, typename ODT>
+ void backward_kernel(const IDT *src_ptr,
+                            ODT *dst_ptr,
+                            size_t volume,
+                            ffStream_t stream);
+} // namespace Internal
+} // namespace Cast
+} // namespace Kernels
+} // namespace FlexFlow
+
+#endif // _FLEXFLOW_OPS_KERNELS_CAST_KERNELS_H
